requirements.txt
Flask==3.1.1
fuzzywuzzy==0.18.0
google_api_python_client==2.176.0
gspread==6.2.1
pandas==2.3.1
pdfplumber==0.11.7
protobuf==6.31.1
